% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-autoencoder.R
\name{CustomAutoencoder}
\alias{CustomAutoencoder}
\title{CustomAutoencoder}
\description{
A common abstract class for custom autoencoder implementations.
}
\section{Arguments}{


\code{network} Ruta network object indicating the desired architecture.

\code{loss} Character scalar (e.g. \code{"mean_squared_error"}) or Ruta loss object.

\code{weight} Weight of the applied penalty.
}

\section{Methods}{


\verb{$new()} Initialize new object

\verb{$set_autoencoder()} Save a Keras model as internal autoencoder.

\verb{$train()} Train the autoencoder with data.

\verb{$encode()} Once trained, encode new data.
}

