% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer.R, R/skaler.R, R/skaler2.R
\name{Scorer}
\alias{Scorer}
\title{Scorer}
\description{
Supervised class overlap rate reduction. This autoencoder uses Fisher's
Discriminant Ratio to penalize the class overlap.

Supervised overlap reduction with Kullback-Leibler divergence penalty

Supervised overlap reduction with Kullback-Leibler divergence penalty
}
\section{Arguments}{


\code{network} Ruta network object indicating the desired architecture.

\code{loss} Character scalar (e.g. \code{"mean_squared_error"}) or Ruta loss object.

\code{weight} Weight of the applied penalty.



\code{network} Ruta network object indicating the desired architecture.

\code{loss} Character scalar (e.g. \code{"mean_squared_error"}) or Ruta loss object.

\code{weight} Weight of the applied penalty.



\code{network} Ruta network object indicating the desired architecture.

\code{loss} Character scalar (e.g. \code{"mean_squared_error"}) or Ruta loss object.

\code{weight} Weight of the applied penalty.
}

\section{Methods}{


\verb{$new()} Initialize new object

\verb{$set_autoencoder()} Save a Keras model as internal autoencoder.

\verb{$train()} Train the autoencoder with data.

\verb{$encode()} Once trained, encode new data.

\verb{$penalty()} Retrieve the Keras tensor which computes the penalty for the loss.



\verb{$new()} Initialize new object

\verb{$set_autoencoder()} Save a Keras model as internal autoencoder.

\verb{$train()} Train the autoencoder with data.

\verb{$encode()} Once trained, encode new data.

\verb{$penalty()} Retrieve the Keras tensor which computes the penalty for the loss.



\verb{$new()} Initialize new object

\verb{$set_autoencoder()} Save a Keras model as internal autoencoder.

\verb{$train()} Train the autoencoder with data.

\verb{$encode()} Once trained, encode new data.

\verb{$penalty()} Retrieve the Keras tensor which computes the penalty for the loss.
}

