% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicer.R
\name{Slicer}
\alias{Slicer}
\title{Slicer}
\description{
Supervised linear classifier error reduction. This autoencoder uses a Support
Vector Machine-based penalty in order to increase class separability.
}
\section{Arguments}{


\code{network} Ruta network object indicating the desired architecture.

\code{loss} Character scalar (e.g. \code{"mean_squared_error"}) or Ruta loss object.

\code{weight} Weight of the applied penalty.
}

\section{Methods}{


\verb{$new()} Initialize new object

\verb{$set_autoencoder()} Save a Keras model as internal autoencoder.

\verb{$train()} Train the autoencoder with data.

\verb{$encode()} Once trained, encode new data.

\verb{$penalty()} Retrieve the Keras tensor which computes the penalty for the
loss.
}

